Create this imgproxy plugin
Why?
====

Why not? I like what imgproxy's doing. I like how well it does the job. There's
no Jekyll plugin to work with it. So let's pull something together to get it
to work.

What?
=====

Add a builder to put together the path based on the options we pass in the tag.
Add a url generator to create the secure url.
Add some tests.
Throw it all in the vendor directory to get it working before packaging it as a gem.

How do we know this works?
==========================

Check out the changes made in `_posts/2022-03-04-weeknotes-for-the-week-ending-mar-04.md`