Fix chicken and egg issue
When deploying to netlify we are working with a completely clean slate.

No files exist.

Until now we had 2 issues depending on when we compiled the css.

Approach 1: compile the css first.

* The css would exist before the files are generated and written.
* But when the final files are copied into `_site` it would then copy the original source css files

Approach 2: compile the css last.

* The html would be generated first, with no css existing.
* Then the css would be generated and written to `_site`
* If the html is written first, and you want to inline the css (which I do)
  then you can't inline it into the html files because the css doesn't exist yet.

Winning approach: compile and copy css first, generate html, copy the css at the end AGAIN.

This way the css exists for the inlining. And we are assured that the final state of the css in
`_site` is the generated, compiled css.